feat: 서랍장 상세 페이지 완성 (#178)
* feat: url마다 BoxButton 추가

* feat: `MoreProductSection`에 페이지 이동 버튼 추가

- 프로덕트는 최대 5개까지만 노출

* feat: SHARE 버튼 클릭시 클립보드에 url 복사

* fix: `productUrls.map()`에 key 추가

* feat: `ServiceDetail`에 YDS Toast 추가

- SHARE 버튼 클릭시 토스트 메시지로 복사된 걸 알림

* refactor: queryClient.invalidateQueries 쿼리키 수정

- productNo 추가
- MoreProductSection은 백엔드 스펙 변경으로 에러 나서 잠시 주석 처리

* refactor: `MoreProductSection`의 data.length를 if문으로 확인

* fix: slice(0,5)를 `MoreProducttSection` 내부로 이동

- useGetProductByProvider 훅은 Provider 페이지에서도 사용하는 훅이라 여기서 자르면 안됨

* fix: `handleClickShare`를 async 함수로 수정

- 클립보드 접근이 프로미스를 반환함

* refactor: `ServiceDetail`의 Category를 상수 파일로 이동

- CategoryObj로 이름 변경
- 직접 값 넣는 게 아니라 CategoryList를 가공해서 생성하도록 수정

* refactor: `MoreProductSection` 컴포넌트 파일 위치 이동

- src/drawer/components/ServiceDetailContents

* refactor: `ServiceDetail`에서 `Carousel` 분리

* refactor: `ServiceDetail`에서 `Description` 분리

* refactor: `ServiceDetail`에서 `ServiceAction` 분리

- 토스트 위치가 이상한데 YDS 구버전이 남아서 그런건지 확인 필요

* refactor: toastProps 타입 단언 -> const assertion

* refactor: `MoreProductSection` return undefined -> return null

* refactor: `MoreProductSection` data에 구조 분해 할당 적용

- providerName prop은 없어도 되는 값이라 삭제함 (훅에서 받는 데이터에도 있음)

* refactor: `useGetProductDetail`에서 useSuspenseQuery를 바로 반환하게 수정
