+++ src/drawer/hooks/useGetProductDetail.ts
@@ -3,7 +3,7 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 import { getProductDetail } from '../apis/getProductDetail';
 
 export const useGetProductDetail = (productNo: number) => {
-  const { data, refetch, ...queryResult } = useSuspenseQuery({
+  return useSuspenseQuery({
     queryKey: ['productDetail', productNo],
     queryFn: () => {
       return getProductDetail(productNo);
@@ -11,6 +11,4 @@ export const useGetProductDetail = (productNo: number) => {
     retry: false,
     select: (data) => data.detail,
   });
-
-  return { product: data, refetch, ...queryResult };
 };
